Reorganise aave like open sidebars
