+++ features/aave/open/sidebars/components/AaveOpenPositionStopLossLambda.tsx
@@ -1,0 +1,6 @@
+import type { OpenAaveStateProps } from 'features/aave/open/sidebars/sidebar.types'
+import React from 'react'
+
+export function AaveOpenPositionStopLossLambda(_props: OpenAaveStateProps) {
+  return <div>new stop loss sidebar</div>
+}
