+++ features/aave/open/sidebars/components/OpenAaveFailureStateView.tsx
@@ -1,0 +1,41 @@
+import type { SidebarSectionProps } from 'components/sidebar/SidebarSection'
+import { ConnectedSidebarSection, StrategyInformationContainer } from 'features/aave/components'
+import type { OpenAaveStateProps } from 'features/aave/open/sidebars/sidebar.types'
+import { useTranslation } from 'next-i18next'
+import React from 'react'
+import { Grid } from 'theme-ui'
+
+export function OpenAaveFailureStateView({ state, send }: OpenAaveStateProps) {
+  const { t } = useTranslation()
+
+  const sidebarSectionProps: SidebarSectionProps = {
+    title: t(state.context.strategyConfig.viewComponents.sidebarTitle),
+    content: (
+      <Grid gap={3}>
+        <StrategyInformationContainer
+          state={state}
+          changeSlippageSource={(from) => {
+            send({ type: 'USE_SLIPPAGE', getSlippageFrom: from })
+          }}
+        />
+      </Grid>
+    ),
+    primaryButton: {
+      isLoading: false,
+      disabled: false,
+      label: t('open-earn.aave.vault-form.retry-btn'),
+      action: () => send({ type: 'RETRY' }),
+    },
+  }
+
+  return (
+    <ConnectedSidebarSection
+      {...sidebarSectionProps}
+      textButton={{
+        label: t('open-earn.aave.vault-form.back-to-editing'),
+        action: () => send('BACK_TO_EDITING'),
+      }}
+      context={state.context}
+    />
+  )
+}
